+++ ISpinBox.h
@@ -10,7 +10,7 @@ public:
     explicit ISpinBox(QWidget *parent = 0);
     virtual ~ISpinBox();
     virtual void setRange(const double &min, const double& max, const double &step) = 0;
-    virtual int value() const = 0;
+    virtual double value() const = 0;
     virtual QWidget *spinWidget() const = 0;
     virtual QWidget *buttunDownWidget() const = 0;
     virtual QWidget *buttonUpWidget() const = 0;
