+++ LCDSpinBox.h
@@ -32,7 +32,7 @@ public:
     virtual ~LCDSpinBox();
 
     virtual void setRange(const double &min, const double &max, const double &step);
-    virtual int value() const;
+    virtual double value() const;
 
     virtual QWidget *spinWidget() const;
     virtual QWidget *buttunDownWidget() const;
