+++ SpinBox.h
@@ -23,7 +23,7 @@ public:
                      QWidget *parent = 0);
 
     virtual void setRange(const int &min, const int& max);
-    virtual int value() const;
+    virtual double value() const;
 
     virtual QWidget *spinWidget() const;
     virtual QWidget *buttunDownWidget() const;
